# backend/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MoneyMap API",
    "version": "0.1.0",
    "description": "Documentação simplificada para os endpoints principais do MoneyMap."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Desenvolvimento" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Notification": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "message": { "type": "string" },
          "type": { "type": "string", "enum": ["system", "transaction", "recurring", "budget", "goal", "reminder"] },
          "severity": { "type": "string", "enum": ["info", "success", "warning", "error"] },
          "actionUrl": { "type": ["string", "null"] },
          "metadata": { "type": ["object", "null"], "additionalProperties": true },
          "read": { "type": "boolean" },
          "readAt": { "type": ["string", "null"], "format": "date-time" },
          "expiresAt": { "type": ["string", "null"], "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "NotificationCreatePayload": {
        "type": "object",
        "required": ["title", "message"],
        "properties": {
          "title": { "type": "string" },
          "message": { "type": "string" },
          "type": { "type": "string", "enum": ["system", "transaction", "recurring", "budget", "goal", "reminder"] },
          "severity": { "type": "string", "enum": ["info", "success", "warning", "error"] },
          "actionUrl": { "type": "string" },
          "metadata": { "type": ["object", "string"], "additionalProperties": true },
          "expiresAt": { "type": "string", "format": "date-time" }
        }
      }
    }
  },
  "paths": {
...
    "/notifications": {
      "get": {
        "tags": ["Notifications"],
        "summary": "Listar notificações",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": { "type": "string", "enum": ["all", "read", "unread"], "default": "all" }
          },
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1 }
          },
          {
            "name": "perPage",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100 }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de notificações do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "notifications": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Notification" }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer" },
                        "page": { "type": "integer" },
                        "perPage": { "type": "integer" },
                        "totalPages": { "type": "integer" },
                        "unreadCount": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Notifications"],
        "summary": "Criar notificação manual",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NotificationCreatePayload" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Notificação criada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "notification": { "$ref": "#/components/schemas/Notification" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/notifications/mark-all-read": {
      "post": {
        "tags": ["Notifications"],
        "summary": "Marcar todas como lidas",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Quantidade marcada como lida",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "updated": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/notifications/{id}/read": {
      "post": {
        "tags": ["Notifications"],
        "summary": "Marcar como lida",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Notificação atualizada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "notification": { "$ref": "#/components/schemas/Notification" }
                  }
                }
              }
            }
          },
          "404": { "description": "Notificação não encontrada" }
        }
      }
    },
    "/notifications/{id}/unread": {
      "post": {
        "tags": ["Notifications"],
        "summary": "Marcar como não lida",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Notificação atualizada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "notification": { "$ref": "#/components/schemas/Notification" }
                  }
                }
              }
            }
          },
          "404": { "description": "Notificação não encontrada" }
        }
      }
    },
    "/notifications/{id}": {
      "delete": {
        "tags": ["Notifications"],
        "summary": "Remover notificação",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "204": { "description": "Removida" },
          "404": { "description": "Notificação não encontrada" }
        }
      }
    },
...
  }
}